fix: check if it is alread set config

+++ zsh/setup_tools/setting-git-config.sh
@@ -5,6 +5,9 @@ printf '\033[33m%s\033[m\n' "setting git config..."
 # Get current config
 USERNAME=`git config --global user.name`
 EMAIL=`git config --global user.email`
+COLOR_UI=`git config --global color.ui`
+CORE_EDITOR=`git config --global core.editor`
+DELETE_MERGED_BRANCH=`git config --global alias.delete-merged-branch`
 
 printf '\033[33m%s\033[m\n' "username"
 if [ "$USERNAME" == '' ]; then
@@ -17,6 +20,15 @@ if [ "$EMAIL" == '' ]; then
     git config --global user.email "$email"
 fi
 
-git config --global color.ui auto
-git config --global core.editor vim
-git config --global alias.delete-merged-branch "!f () { git checkout $1; git branch --merged|egrep -v '\*|develop|master'|xargs git branch -d; };f"
+printf '\033[33m%s\033[m\n' "core.ui"
+if [ "$COLOR_UI" == '' ]; then
+	git config --global color.ui auto
+fi
+printf '\033[33m%s\033[m\n' "core.editor"
+if [ "$CORE_EDITOR" == '' ]; then
+	git config --global core.editor vim
+fi
+printf '\033[33m%s\033[m\n' "alias"
+if [ "$DELETE_MERGED_BRANCH" == '' ]; then
+	git config --global alias.delete-merged-branch "!f () { git checkout $1; git branch --merged|egrep -v '\*|develop|master'|xargs git branch -d; };f"
+fi
